pyproject.toml
[tool.poetry]
name = "course-beta"
version = "0.1.0"
description = ""
authors = ["<Mary> <<kass.o@yandex.ru>>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.7.1"
black = "^23.11.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
yfinance = "^0.2.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 119

[tool.mypy]
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
ignore_missing_imports = true

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''
